Added new functions to render tasks and clear unused DOM elements
I have added two new functions to the project that will help with rendering tasks and managing DOM elements.

The first function is called renderTasks(), which takes an array of tasks as input and renders them into the DOM. This function loops through each task in the array, creates a new HTML element for it, and appends it to the appropriate section in the DOM. This will make it easier to display all the tasks in the application and keep them organized.

The second function is called clearUnusedElements(), which clears any DOM elements that are no longer in use. This function takes an array of elements that should be kept and removes all other elements from the DOM. This will help keep the DOM clean and prevent memory leaks, which can slow down the application over time.

Together, these two functions will help make the application more efficient and easier to use by keeping the DOM organized and free from unnecessary elements.