FEAT: Respuesta a pregunta No. 7
This code will select the following columns from the car_brands and car_models tables:

car_brands.brand_name: The name of the car brand.
car_models.model: The name of the car model.
car_models.sale_price: The sale price of the car.
The code will then join the two tables on the id column. This will ensure that the selected rows are from the same car brand.

Finally, the code will filter the results to only include cars that have a sale price greater than $20,000.

The output of this code will be a list of rows that contain the following information:

The name of the car brand.
The name of the car model.
The sale price of the car.
This code can be used to find all cars that are sold for more than $20,000. This information can be used for marketing purposes, to identify popular car models, or to track the price of cars over time.