FEAT: Answer No. 8
This code will select the following information from the database:

The name of the car dealer
The name of the customer
The model of the car
The code will only select rows where the order state is "pending".

Here is a breakdown of the code:

The SELECT statement specifies the columns that we want to select.
The FROM clause specifies the tables that we want to join.
The INNER JOIN clauses specify how we want to join the tables.
The WHERE clause specifies the condition that we want to filter the rows.
I hope this helps! Let me know if you have any other questions.