FEAT: Get customers with orders over $300
Changes:
This commit adds a query to get the names, emails, and addresses of customers who have placed orders over $300.
Testing:
The query was tested by running it on a database with a few orders over $300. The query successfully returned the names, emails, and addresses of the customers who had placed those orders.

Deployment:
The query was deployed to the database by running the following command:

Code snippet
mysql> use my_database
mysql> source get_customers_with_orders_over_300.sql
Use code with caution. Learn more
The query is now available to be used in other queries or applications.

I hope this helps!