Create trigger to update order date
Changes:
This commit adds a trigger to update the order date whenever an order is updated.
Testing:
The trigger was tested by updating an order in the database. The trigger successfully updated the order date to the current date and time.

Deployment:
The trigger was deployed to the database by running the following command:
mysql> use my_database
mysql> drop trigger if exists update_order;
mysql> source update_order.sql