Update order total
Changes:
This commit updates the order total for order ID 1 to $250.00.
Testing:
The update was tested by running it on a database with an order ID of 1. The update successfully updated the order total to $250.00.

Deployment:
The update was deployed to the database by running the following command:

Code snippet
mysql> use my_database
mysql> source update_order_total.sql
Use code with caution. Learn more
The update is now available to be used in other queries or applications.

I hope this helps!