FEAT: Create data table
Created a data table using PHP to display data from a database. The table is formatted using CSS and can be easily customized.

Changes
Created a new PHP file to create the data table.
Added CSS to format the table.
Added JavaScript to make the table interactive.
Testing
The data table has been tested on Chrome, Firefox, and Safari. It works correctly on all three browsers.

Next steps
Add more features to the data table, such as sorting and filtering.
Make the data table responsive so that it works on mobile devices.