Initial commit of new PHP project
A index.php file that serves as the entry point for the application.
A assets folder that contains all the css, img and JavaScript needed for the project.
A src folder that contains the data base connection, the models and the views.

This is just the initial commit for the project. More commits will be added as the project progresses.