FEAT: Error Messages
In this commit, I refactored the code to improve the validation and sanitization of form fields. Here are the changes made:

Moved the code inside an if-else block for better organization.
Replaced the Spanish echo statements with comments to indicate that they should be removed.
Sanitized the firstName, lastName1, lastName2, and email variables using the sanitizePhrase and sanitizeEmail functions respectively.
Added a comment to indicate the output of the variables.
Please review and merge this commit as appropriate