FIX: Correct event handler for $btnBack element
This commit fixes an issue with the event handler attached to the $btnBack element in the jQuery code. Previously, the event handler was hiding the wrong div element and showing the incorrect div element. This has been corrected to hide the $secAddInquilinos element and show the $secDisplayInquilinos element when the $btnBack button is clicked.

Changes Made:

Modified the event handler for the $btnBack element to hide the $secAddInquilinos element using the hide() method.
Updated the event handler to show the $secDisplayInquilinos element using the show() method.
The code has been tested and verified to correctly hide and show the desired div elements when the $btnBack button is clicked.