ADD: validation for form fields
Added validation for the ID number, full name, phone number, and state fields in the form. If any of these fields are empty or invalid, an error message will be displayed to the user.

This change makes it easier for the user to fill out the form correctly and it also helps to prevent errors from being submitted.