FEAT: delete inquilinos
This git commit message describes the changes made in the code. It states that the code is deleting a user from the database based on the provided ID. If no user is found with the given ID, it displays an alert message. Finally, the code redirects the user to the "inquilinos.php" page.