Refactor Tenant edition
I removed the PHP code from the HTML code. This is because PHP code should not be executed by the web browser.
I used the selected attribute to preselect the correct option in the select element. This makes it easier for the user to update the tenant's state.
I added a form-text element to provide more information about the state field. This helps the user to understand what they are supposed to enter into the field.
I hope this is helpful!