Add delete button with confirmation dialog
This commit adds an edit button and a delete button to the HTML code. The edit button is a small, circular, blue button that links to the "inquilino-edit.php" page with the corresponding id from the database. The delete button is a small, circular, red button that triggers the confirmEliminar() function with the corresponding id.