Add PHP code for activity page
This commit adds the PHP code for the activity page (actividad.php). The code begins by declaring the necessary variables and including the database connection file (connection-db.model.php). The connection to the database is established using the openConnection() function.

If the connection is successful, the code proceeds to select all items from the visitas table using a SQL query. The query is executed using the $connection object, and the result is stored in the $result variable.

Next, the code checks for any errors in the query execution. If an error occurs, the script will terminate and display an error message. Otherwise, the number of rows returned by the query is assigned to the $activityNumber variable using the num_rows property of the $result object.

Finally, the code includes the view file (actividad.view.php) that will display the activity page.

Note: This commit does not include any additional functionality or styling. It solely focuses on retrieving and displaying the number of rows from the visitas table.

Task ID: #activity-page
Issue: N/A