Update section-display-guests.html to add a table of visits
Added a table of visits to the section-display-guests.html file.
The table includes the following columns:
Visit ID
Visitor ID
Destination ID
Time of entry
Observations
Buttons to edit and delete the visit
The table is styled using Bootstrap.
This commit updates the section-display-guests.html file to include a table of visits. The table includes the following columns:

Visit ID
Visitor ID
Destination ID
Time of entry
Observations
Buttons to edit and delete the visit
The table is styled using Bootstrap.

This commit was made on 2023-06-22 at 12:12 PM PST by Bard.

I hope this helps!