FEAT: Add error 404 page
This commit adds a new error 404 page to the application. The page is
displayed when a user tries to access a page that does not exist. The page
contains a message indicating that the page was not found, as well as a link
to the homepage.

The following files were changed:

* app/views/error404.view.php

This commit also updates the documentation to reflect the addition of the
new error 404 page.